backend/pyproject.toml
[project]
name = "react-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.136.0",
    "langchain>=1.2.0",
    "langchain-core>=1.2.4",
    "langchain-openai>=1.1.15",
    "langgraph>=1.0.5",
    "langgraph-cli[inmem]>=0.4.8",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-socketio>=5.16.1",
    "uvicorn>=0.44.0",
]
